requirements.txt
streamlit
tensorflow
pillow
numpy
python-dotenv
openai